Fix crash when NFC process killed during BT Headset handover
The caused of this issue is due to illegal argument exception
thrown in BT Peripheral Handover. The previously registered
receiver no longer existed in Activity Manager when the NFC
process was killed. Because of that the unregister function
in BT Peripheral Handover during completion throws an
exception.

During the death of NFC process, it will unbind itself from
Handover service, then eventually the Handover process will be
destroyed. In order to fix this issue, an ongoing BT Peripheral
Handover should be stopped during unbinding process.

Change-Id: I13b6d313d123af8de6809be848519ae2f5644f22